requirements.txt
ExtractTable==2.1.2
Pillow==8.2.0
